+++ Changelog.md
@@ -1,5 +1,6 @@
 Version 0.3.1 (Unrelease)
 =========================
+- Implement `AsMut<Span>` for `AutoFinishingSpan`.
 - Implement `AsMut<Span>` for `Span`.
 - Utility trait to fail spans on `Err`.